+++ src/components/Window/Canvas/Canvas.module.scss
